requirements.txt
astroid==2.3.3
certifi==2019.3.9
chardet==3.0.4
Click==7.0
idna==2.8
isort==4.3.21
Jinja2==2.10.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pip==19.3.1
pylint==2.4.4
python-gitlab==1.8.0
requests==2.22.0
setuptools==41.6.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
wheel==0.33.6
wrapt==1.11.2
